Match output to input and clarify build command
The Dockerfile listed above specifically uses package.json, but the
output listed for the build lists package*, which is inconsistent. The
React example later in the doc uses package*, which would be consistent
with the output, but this should fix the mismatch for the beginner going
through the tutorial for the first time.

Additionally, the sentence "Build the Docker image now using 'docker
build' again" might be confusing to a beginner. I specifically tested
running this exact command and it gives an error message. Specifically
using the exact command needed to build this target shouldn't take up
too much space.