Fix #76
If the user runs that command "from the source code folder", i.e.
`getting-started`, docker will bind-mount getting-started to the
container's `/app`, which leads to error
`error Couldn't find a package.json file in "/app"`. It needs to bind
`getting-started/app` instead.
So, be explicit to the reader to be on the app source's directory
`getting-started/app` before running the command.